fix(lobby): get/create/join/leave coinche game rooms
