+++ client/package.json
@@ -16,10 +16,12 @@
     "node": "12"
   },
   "dependencies": {
+    "axios": "0.19.2",
     "boardgame.io": "0.39.10",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "react-router-dom": "5.1.2"
+    "react-router-dom": "5.2.0",
+    "swr": "0.2.0"
   },
   "devDependencies": {
     "@types/jest": "25.2.1",
@@ -31,7 +33,7 @@
     "lint-staged": "10.2.2",
     "npm-run-all": "4.1.5",
     "react-scripts": "3.4.1",
-    "typescript": "3.8.3"
+    "typescript": "3.9.2"
   },
   "browserslist": {
     "production": [
