+++ client/src/client/lobby/Lobby.css
@@ -1,0 +1,4 @@
+/**
+ */
+.lobby {
+}
