+++ client/src/client/lobby/i18n/fr.ts
@@ -1,0 +1,8 @@
+import {I18n} from './index';
+
+export const fr: I18n = {
+  createRoom: 'Créer une nouvelle partie',
+  joinRoom: 'Rejoindre',
+  leaveRoom: 'Quitter',
+  waitingForPlayer: 'Place disponible',
+};
