+++ client/src/client/lobby/i18n/index.ts
@@ -1,0 +1,8 @@
+export type I18n = {
+  createRoom: string;
+  joinRoom: string;
+  leaveRoom: string;
+  waitingForPlayer: string;
+};
+
+export { fr } from './fr';
