+++ client/src/shared/index.ts
@@ -1,0 +1,5 @@
+export type LanguageCode = 'fr';
+
+export enum GameName {
+  Coinche = 'coinche'
+}
