+++ server/package.json
@@ -37,7 +37,7 @@
     "supertest": "4.0.2",
     "ts-jest": "25.4.0",
     "ts-node-dev": "1.0.0-pre.44",
-    "typescript": "3.8.3"
+    "typescript": "3.9.2"
   },
   "husky": {
     "hooks": {
